intersection-observer/zero-area-element-hidden.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
pre, #log {
  position: absolute;
  top: 0;
  left: 200px;
}
#target {
  width: 0px;
  height: 0px;
  position: fixed;
  top: -1000px;
}
</style>

<div id='target'></div>

<script>
var entries = [];

runTestCycle(function() {
  assert_equals(window.innerWidth, 800, "Window must be 800 pixels wide.");
  assert_equals(window.innerHeight, 600, "Window must be 600 pixels high.");

  var target = document.getElementById('target');
  assert_true(!!target, "target exists");
  var observer = new IntersectionObserver(function(changes) {
    entries = entries.concat(changes)
  });
  observer.observe(target);
  entries = entries.concat(observer.takeRecords());
  assert_equals(entries.length, 0, "No initial notifications.");
  runTestCycle(step0, "First rAF.");
}, "A zero-area hidden target should not be intersecting.");

function step0() {
  checkLastEntry(entries, 0, [8, 8, -1000, -1000, 0, 0, 0, 0, 0, 800, 0, 600, false]);
}
</script>
